requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
antiorm==1.2.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
db==0.1.1
db-sqlite3==0.0.1
Deprecated==1.2.18
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
